service/requirements.txt
fastapi~=0.89.1
pydantic~=1.10.4
motor~=3.1.1
uvicorn
py-automapper